Make your list sortable
Making the list sortable is easy using the jQuery-UI Sortable plugin.

Make the library available to your application by adding the `jquery-ui-rails` gem to your `Gemfile`. Then run `bundle install` to install the gem. Finally, you need to require jquery-ui in your javascript manifest to actually include the library in the HTML.

To make the list sortable, we need to select it using jQuery. Let's add an `id` to the `ul` to make this easier.

Then, create a new Javascript file to add our sorting logic. It will be loaded automatically because of the `require_tree .` line in your manifest file. We put all of our jQuery code inside a `$(document).ready` function to make sure that it isn't run until the document is loaded.

You'll need to restart your Rails server for the new gem to be picked up. When you reload your list, you should now find that you can drag and drop them. Awesome! Of course, if you reload the page you'll find that the order has reset. Don't worry, we'll work on saving the position changes to the database next.