Add an action to sort the FAQs
We'll be making an AJAX request to update the FAQs' positions whenever they are sorted. Let's create a controller action that will update the positions.

We need to update the `position` attribute of many FAQs at once. If the last one gets sorted to the top, it will get position 1, and all the others will have their position incremented by one. So an `update` action doesn't really make sense. Let's create an action called `sort` instead, which will handle updating the sorting of all FAQs.

The end-point for this action should be `POST /faqs/sort`, so we add it as a collection route to the FAQs resource in `routes.rb`.

This action will expect the `faq` param to be a list of IDs of FAQs in the desired order. The POST parameters convention for a list is a bit funny looking; it would look something like `faq[]=3&faq[]=1&faq[]=2`. Rails will turn this into a handy array for us.

The rest is easy. We loop over this array and update the position of each FAQ we see with its position in the array.

Finally, we render a simple 200 status code to signify that the sorting happened successfully.