Post the positions to the server using AJAX
Finally, we need our Javascript code to send the new positions to the server whenever the sorting is updated.

If you pass a function to the `update` option to `sortable`, it will get called whenever there is an update to the sorting.

We'd like to send a POST request to `/faqs/sort` with a list of the FAQs' `id`s in the desired order. `sortable` makes this really easy. All we have to do is add an `id` attribute to each `li` that contains the FAQ's `id`, and then calling `sortable('serialize')` will give us the list exactly as we want.

Finally, we post the list to our server using jQuery.

Try it out. Sort the list, and then refresh the page. The changes you make should persist. You, my friend, just created a sortable list in Rails! Now don't get cocky, kid.